+++ docs/release_notes.rst
@@ -6,6 +6,7 @@ Release Notes
 
   * Added the :meth:`~dicom_parser.header.Header.keyword_contains` method to
     easily look for data elements using a substring of the keyword.
+  * Increased tests coverage.
 
 0.1.6
 -----